New logic to present the standings in descending order by points, then by goal differe then by goals scored.New span errors for creating a match.Additional logic to calculate the points and e.t.c after the match is finished.

+++ src/main/java/firstwebjavaproject/javawebproject/model/Match.java
@@ -84,7 +84,7 @@ public class Match {
         return homeTeamScore;
     }
 
-    public Integer getLeagueId(){
+    public Long getLeagueId(){
         return this.league.getId();
     }
